app/requirements.txt
Django==6.0.3
psycopg[binary]==3.2.6
celery==5.6.2
redis==6.1.0
python-dotenv==1.1.0
gunicorn==23.0.0
whitenoise==6.9.0
argon2-cffi==25.1.0
Pillow==11.2.1
